+++ protected/humhub/modules/installer/controllers/SetupController.php
@@ -10,6 +10,7 @@ namespace humhub\modules\installer\controllers;
 
 use humhub\components\access\ControllerAccess;
 use humhub\components\Controller;
+use humhub\helpers\ConfigHelper;
 use humhub\modules\admin\widgets\PrerequisitesList;
 use humhub\modules\installer\forms\DatabaseForm;
 use humhub\modules\installer\libs\DynamicConfig;
@@ -66,8 +67,11 @@ class SetupController extends Controller
 
         $dynamicConfig = new DynamicConfig();
         $model = new DatabaseForm();
+        $model->autoLoad();
 
-        if (($model->autoLoad() || $model->load(Yii::$app->request->post())) && $model->validate()) {
+        $postLoaded = $model->load(Yii::$app->request->post());
+
+        if (($this->module->enableAutoSetup || $postLoaded) && $model->validate()) {
             try {
                 if ($model->create) {
                     /** @var yii\db\Connection $temporaryConnection */
@@ -83,14 +87,18 @@ class SetupController extends Controller
                     }
                 }
 
-                $dbConfig = $model->getDbConfigAsArray(true);
-
+                $dbConfig = $model->getDbConfigAsArray();
                 /** @var yii\db\Connection $temporaryConnection */
                 $temporaryConnection = Yii::createObject($dbConfig);
                 // Try access to the given database
                 $temporaryConnection->open();
 
-                $dynamicConfig->content['components']['db'] = $dbConfig;
+                if ($postLoaded) {
+                    $dynamicConfig->content['components']['db'] = $model->getIncompleteFixedConfig();
+                } else {
+                    $dynamicConfig->content = [];
+                }
+
                 $dynamicConfig->save();
 
                 return $this->redirect(['migrate']);
@@ -99,8 +107,14 @@ class SetupController extends Controller
             }
         }
 
-        $model->password = '';
-        return $this->render('database', ['model' => $model, 'errorMessage' => $errorMessage]);
+        if (!$model->isFixed('password')) {
+            $model->password = '';
+        }
+
+        return $this->render('database', [
+            'model' => $model,
+            'errorMessage' => $errorMessage
+        ]);
     }
 
 
@@ -119,7 +133,7 @@ class SetupController extends Controller
      */
     public function actionCron()
     {
-        if ($this->module->enableAutoSetup) {
+        if ($this->module->enableAutoSetup || filter_var($_ENV['HUMHUB_DOCKER'] ?? false, FILTER_VALIDATE_BOOLEAN)) {
             return $this->redirect(['finalize']);
         }
 
@@ -141,7 +155,13 @@ class SetupController extends Controller
      */
     public function actionPrettyUrls()
     {
-        if ($this->module->enableAutoSetup) {
+        if (
+            $this->module->enableAutoSetup ||
+            ConfigHelper::instance()->get(
+                'components.urlManager.enablePrettyUrl',
+                ConfigHelper::SET_COMMON | ConfigHelper::SET_ENV
+            )
+        ) {
             return $this->redirect(['finalize']);
         }